Move clipboard functionality into own component
